Refactor web API services and remove WebAPIBase class Refactored MicroMController in smaller controllers Refactored WebAPIBase in  differnt services  Removed uniqueConstraint in menu path due to path field size, moved the logic for uniqueness to nmi_update
Updated method signatures across multiple classes to use IWebAPIServices instead of IMicroMWebAPI, ensuring a consistent interface. Modified EntityBase methods for alignment with the new service interface. Introduced or updated controllers for authentication and entity handling. Added DependencyInjectionDebug for logging DI events and enhanced MicroMAppConfigurationProvider with logging for configuration management. Simplified PublicEndpointSecurityRecord for better maintainability. Updated SQL scripts for MicromMenusItems to sync with application logic. Overall, these changes improve structure, maintainability, and functionality of the application.
