+++ MicroM/core/DataDictionary/Entities/MicromUsers/MicromUsersGroups.cs
@@ -4,61 +4,58 @@ using MicroM.Core;
 using MicroM.Data;
 using MicroM.Web.Services;
 
-namespace MicroM.DataDictionary
+namespace MicroM.DataDictionary;
+
+
+public class MicromUsersGroupsDef : EntityDefinition
 {
+    public MicromUsersGroupsDef() : base("mug", nameof(MicromUsersGroups)) { SQLCreationOptions = SQLCreationOptionsMetadata.WithIUpdate; }
 
-    public class MicromUsersGroupsDef : EntityDefinition
+    public readonly Column<string> c_user_group_id = Column<string>.PK(autonum: true);
+    public readonly Column<string> vc_user_group_name = Column<string>.Text();
+
+    public readonly Column<string[]?> vc_group_members = Column<string[]?>.Text(size: 0, fake: true, isArray: true, nullable: true);
+
+    public readonly ViewDefinition mug_brwStandard = new(nameof(c_user_group_id));
+
+    public readonly ProcedureDefinition mug_GetAllGroupsAllowedRoutes = new(readonly_locks: true);
+
+    public readonly EntityUniqueConstraint UNGroupName = new(keys: [nameof(vc_user_group_name)]);
+
+    public readonly EntityForeignKey<MicromUsers, MicromUsersGroups> FKUsers = new(fake: true);
+
+}
+
+public class MicromUsersGroups : Entity<MicromUsersGroupsDef>
+{
+    public MicromUsersGroups() : base() { }
+    public MicromUsersGroups(IEntityClient ec, IMicroMEncryption? encryptor = null) : base(ec, encryptor) { }
+
+    public override async Task<DBStatusResult> InsertData(CancellationToken ct, bool throw_dbstat_exception = false, MicroMOptions? options = null, Dictionary<string, object>? claims = null, IWebAPIServices? api = null, string? app_id = null)
     {
-        public MicromUsersGroupsDef() : base("mug", nameof(MicromUsersGroups)) { SQLCreationOptions = SQLCreationOptionsMetadata.WithIUpdate; }
+        var result = await base.InsertData(ct, throw_dbstat_exception, options, claims, api, app_id);
 
-        public readonly Column<string> c_user_group_id = Column<string>.PK(autonum: true);
-        public readonly Column<string> vc_user_group_name = Column<string>.Text();
+        if (api != null && !string.IsNullOrEmpty(app_id)) await api.securityService.RefreshGroupsSecurityRecords(app_id, ct);
 
-        public readonly Column<string[]?> vc_group_members = Column<string[]?>.Text(size: 0, fake: true, isArray: true, nullable: true);
-
-        public readonly ViewDefinition mug_brwStandard = new(nameof(c_user_group_id));
-
-        public readonly ProcedureDefinition mug_GetAllGroupsAllowedRoutes = new(readonly_locks: true);
-
-        public readonly EntityUniqueConstraint UNGroupName = new(keys: [nameof(vc_user_group_name)]);
-
-        public readonly EntityForeignKey<MicromUsers, MicromUsersGroups> FKUsers = new(fake: true);
-
+        return result;
     }
 
-    public class MicromUsersGroups : Entity<MicromUsersGroupsDef>
+    public override async Task<DBStatusResult> UpdateData(CancellationToken ct, bool throw_dbstat_exception = false, MicroMOptions? options = null, Dictionary<string, object>? server_claims = null, IWebAPIServices? api = null, string? app_id = null)
     {
-        public MicromUsersGroups() : base() { }
-        public MicromUsersGroups(IEntityClient ec, IMicroMEncryption? encryptor = null) : base(ec, encryptor) { }
+        var result = await base.UpdateData(ct, throw_dbstat_exception, options, server_claims, api, app_id);
 
-        public override async Task<DBStatusResult> InsertData(CancellationToken ct, bool throw_dbstat_exception = false, MicroMOptions? options = null, Dictionary<string, object>? claims = null, IMicroMWebAPI? api = null, string? app_id = null)
-        {
-            var result = await base.InsertData(ct, throw_dbstat_exception, options, claims, api, app_id);
+        if (api != null && !string.IsNullOrEmpty(app_id)) await api.securityService.RefreshGroupsSecurityRecords(app_id, ct);
 
-            if (api != null && !string.IsNullOrEmpty(app_id)) await api.SecurityService.RefreshGroupsSecurityRecords(app_id, ct);
-
-            return result;
-        }
-
-        public override async Task<DBStatusResult> UpdateData(CancellationToken ct, bool throw_dbstat_exception = false, MicroMOptions? options = null, Dictionary<string, object>? server_claims = null, IMicroMWebAPI? api = null, string? app_id = null)
-        {
-            var result = await base.UpdateData(ct, throw_dbstat_exception, options, server_claims, api, app_id);
-
-            if (api != null && !string.IsNullOrEmpty(app_id)) await api.SecurityService.RefreshGroupsSecurityRecords(app_id, ct);
-
-            return result;
-        }
-
-        public override async Task<DBStatusResult> DeleteData(CancellationToken ct, bool throw_dbstat_exception = false, MicroMOptions? options = null, Dictionary<string, object>? server_claims = null, IMicroMWebAPI? api = null, string? app_id = null)
-        {
-            var result = await base.DeleteData(ct, throw_dbstat_exception, options, server_claims, api);
-
-            if (api != null && !string.IsNullOrEmpty(app_id)) await api.SecurityService.RefreshGroupsSecurityRecords(app_id, ct);
-
-            return result;
-        }
-
+        return result;
     }
 
+    public override async Task<DBStatusResult> DeleteData(CancellationToken ct, bool throw_dbstat_exception = false, MicroMOptions? options = null, Dictionary<string, object>? server_claims = null, IWebAPIServices? api = null, string? app_id = null)
+    {
+        var result = await base.DeleteData(ct, throw_dbstat_exception, options, server_claims, api);
+
+        if (api != null && !string.IsNullOrEmpty(app_id)) await api.securityService.RefreshGroupsSecurityRecords(app_id, ct);
+
+        return result;
+    }
 
 }
